Generic fixes

+++ q1/q1.c
@@ -7,17 +7,17 @@
 
 #define buffLen 60000
 
-int min(int a, int b)
+long long min(long long a, long long b)
 {
     return a < b ? a : b;
 }
 
-int max(int a, int b)
+long long max(long long a, long long b)
 {
     return a > b ? a : b;
 }
 
-void main(int argc, char **argv)
+void main(long long argc, char **argv)
 {
     if (argc != 2)
     {
@@ -27,17 +27,17 @@ void main(int argc, char **argv)
         return;
     }
     char buff[buffLen];
-    int input = open(argv[1], O_RDONLY);
+    long long input = open(argv[1], O_RDONLY);
     mkdir("./Assignment", 0777);
     const char *outputPath = "./Assignment/1_";
     char outputFileName[strlen(argv[1]) + strlen(outputPath) + 1];
     strcpy(outputFileName, outputPath);
     strcat(outputFileName, argv[1]);
 
-    int output = open(outputFileName, O_WRONLY | O_CREAT | O_TRUNC, 0666);
+    long long output = open(outputFileName, O_WRONLY | O_CREAT | O_TRUNC, 0666);
 
-    int inputLength = 0;
-    int curLen = 0;
+    long long inputLength = 0;
+    long long curLen = 0;
     while ((curLen = read(input, buff, buffLen)) > 0)
     {
         inputLength += curLen;
@@ -45,14 +45,14 @@ void main(int argc, char **argv)
     // printf("%d\n",inputLength);
 
     lseek(input, -min(buffLen, inputLength), SEEK_END);
-    int outputLength = 0;
-    int endLength = -1;
+    long long outputLength = 0;
+    long long endLength = -1;
     while ((curLen = read(input, buff, buffLen)) > 0 && outputLength < inputLength)
     {
         curLen = endLength == -1 ? curLen : endLength;
         lseek(input, -curLen, SEEK_CUR);
         // printf("\nWriting %d bytes starting with %c\n", curLen, buff[0]);
-        for (int i = 0; i < curLen / 2; i++)
+        for (long long i = 0; i < curLen / 2; i++)
         {
             char temp = buff[i];
             buff[i] = buff[curLen - i - 1];
